chore: move trace link to a dedicated module
