+++ public/app/features/explore/TraceView/TraceView.tsx
@@ -39,7 +39,8 @@ import type TTraceTimeline from './components/types/TTraceTimeline';
 import { type SpanLinkFunc } from './components/types/links';
 import { type Trace } from './components/types/trace';
 import { isSummarySpan } from './components/utils/summary-span';
-import { createSpanLinkFactory, createTraceLogsLink } from './createSpanLink';
+import { createSpanLinkFactory } from './createSpanLink';
+import { createTraceLogsLink } from './createTraceLink';
 import { useChildrenState } from './useChildrenState';
 import { useDetailState } from './useDetailState';
 import { useHoverIndentGuide } from './useHoverIndentGuide';
